updated nn_server installing scripts, updated nn_server folder structure, added new neural networks to nn_server.py
